+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `offline` module to download a JSON file version of the database
 - Refactor functions useful in both online and offline mode into a `utils` module
 - Create an `OffFlexilims` class to handle offline mode
+- `Flexilims.Session` can be given a token at creation
 
 # v0.5